requirements.txt
requests~=2.32.3
openai~=1.40.6
python-dotenv~=1.0.1
pydantic~=2.8.2
moviepy~=2.0.0.dev2
python-telegram-bot~=21.4